`clang-tidy` everything, everywhere, all at once (#489)
* Enable clang-tidy on examples and tests
* clang-tidy 18 fixes

+++ src/core/data/detail/array_kind.h
@@ -16,7 +16,7 @@
 
 namespace legate::detail {
 
-enum class ArrayKind : std::int32_t {
+enum class ArrayKind : std::uint8_t {
   BASE   = 0,
   LIST   = 1,
   STRUCT = 2,
